refactor: split views/redux/info/quests into focused modules
Converts the 754-line quests.ts into a views/redux/info/quests/ directory:
types / time (tanakalendar + reset-boundary helpers) / goal-matching
(satisfyGoal, satisfyShip) / records (record lifecycle + progress update) /
persistence (quest-tracking CSON writer, FileWriter singleton) / index
(the RTK slice, goal loading, schedualDailyRefresh, and explicit re-export
of the previous public surface).

The module specifier 'views/redux/info/quests' resolves to the directory
unchanged for info/index.ts, create-store.ts, task-panel.tsx, the spec,
and any plugins.

Pure move, with one cleanup: the local `declare const APPDATA_PATH/ROOT`
lines are dropped in favor of the global declarations that now exist
(shims/global.d.ts / lib/globals.d.ts).
